Time: 0 ms (100.00%) | Memory: 9.1 MB (90.92%) - LeetSync

+++ 377-combination-sum-iv/combination-sum-iv.cpp
@@ -27,7 +27,7 @@ public:
     int combinationSum4(vector<int>& nums, int target) {
         int n=nums.size();
         sort(nums.begin(),nums.end());
-        vector<long long>dp(target+1,0);
+        vector<int>dp(target+1,0);
         dp[0]=1;
         for(int i=1;i<=target;i++){
         unsigned long long ans=0;
